+++ src/ShaderPlayground.Core/CompilerNames.cs
@@ -14,5 +14,6 @@
         public const string Slang = "slang";
         public const string GlslOptimizer = "glsl-optimizer";
         public const string Hlsl2Glsl = "hlsl2glsl";
+        public const string HlslParser = "hlslparser";
     }
 }